Update the database states on the basis of stdout and response code

+++ test_db/db.sh
@@ -12,4 +12,13 @@ else
 fi
 
 $db_cmd < $tmp
-rm -f $tmp
+
+if [ $? -eq 0 ]
+then
+  echo "success"
+  rm -f $tmp
+else
+  echo "failure" >&2
+fi
+
+
